memory: brcmstb: dpfe: skip downloading firmware for secure mode
If the system is running in secure mode, Linux is not able to access
the DCPU registers to download the firmware. Attempting to access the
memory will result in a "target abort" error.

Also, when the system is running in secure mode, the DPFE firmware has
already been loaded by the boot loader, which means that Linux can
simply skip downloading the firmware in this case.

Therefore, the driver now checks if the DCPU is already running and, if
so, whether it can process commands. If the DCPU processes commands
successfully, the driver assumes the system is running in secure mode
and simply skips the firmware download step.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>