soc: brcmstb: dpfe: use byte 3 of registers MR4-MR8
For dual-rank LPDDR4, result data is placed in byte 0 and byte 3 of the
corresponding MR register by the firmware. This was supposed to be
done for single-rank RAM, as well. Due to a firmware bug, result
values are only placed in byte 3 of the corresponding MR register.

Since byte 3 works for single-rank and dual-rank setups, we change the
Linux driver to always use byte 3, thus returning the correct value in
either case.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>